requirements.txt
pathway
google-generativeai
python-dotenv
requests
streamlit
scikit-learn
numpy
